Fixed minor bugs.

+++ MaliciousApplicationDetector/dataset.py
@@ -22,6 +22,9 @@ from webgraphic.webgraphic import *
 __all__ = ['dataset']
 
 
+ROOT = os.path.split(os.path.abspath(__file__))[0]
+
+
 FLOW_DICT = {
     'Browser_PC' : lambda stream: stream.GetBrowserGroup_PC(),
     'Background_PC' : lambda stream: stream.GetBackgroudGroup_PC(),
@@ -69,6 +72,7 @@ def dataset(*args, mode):
     # check status
     while any(_worker_alive):
         time.sleep(random.randint(0, dt.datetime.now().second))
+    print(f'Dataset ready @ {make_path(f'dataset')}')
 
     # dump index.json
     return make_index()
@@ -77,28 +81,39 @@ def dataset(*args, mode):
 def worker(path, *, mode, _count=0):
     """Prepare dataset for CNN."""
     global _worker_alive
+    print(f'Worker No.{_count+1} @mode_{mode} on {path}')
 
     # extract name
     root, file = os.path.split(path)
     name, ext = os.path.splitext(file)
 
     # duplicate PCAP file
-    while pathlib.Path(f'./stream/{name}').exists():
+    while pathlib.Path(make_path(f'stream/{name}')).exists():
         name = f'{name}_{int(time.time())}'
-    pathlib.Path(f'./stream/{name}').mkdir(parents=True, exist_ok=True)
-    shutil.copy(path, f'./stream/{name}/{name}.pcap')
+    pathlib.Path(make_path(f'stream/{name}')).mkdir(parents=True, exist_ok=True)
+    shutil.copy(path, make_path(f'stream/{name}/{name}.pcap'))
 
     # make files
-    sdict = make_steam(name, mode=mode)         # make stream
-    os.kill(os.getppid(), signal.SIGUSR1)       # send signal
-    index = make_dataset(name, labels=sdict)    # make dataset
+    sdict = make_steam(name, mode=mode)             # make stream
+    if mode != 2:
+        os.kill(os.getppid(), signal.SIGUSR1)       # send signal
+    index = make_dataset(name, sdict, mode=mode)    # make dataset
 
     # aftermath
-    if path != f'./stream/{name}/{name}.pcap':
-        os.remove(f'./stream/{name}/{name}.pcap')
+    if path != make_path(f'stream/{name}/{name}.pcap'):
+        os.remove(make_path(f'stream/{name}/{name}.pcap'))
 
     # update status
     _worker_alive[_count] = False
+
+    # send signal
+    if mode == 2:
+        os.kill(os.getppid(), signal.SIGUSR1)
+
+
+def make_path(path):
+    """Make path."""
+    return os.path.join(ROOT, path)
 
 
 def make_worker(signum=None, stack=None):
@@ -133,19 +148,22 @@ def make_steam(name, *, mode):
         * dict -- dataset labels
 
     """
+    print(f'Start labeling for {name}...')
+
     # Web Graphic
     builder = webgraphic()
-    builder.read_in(f'./stream/{name}/{name}.pcap')
+    builder.read_in(make_path(f'stream/{name}/{name}.pcap'))
     IPS = builder.GetIPS()
 
     # Stream Manager
     stream = StreamManager(f'{name}.pcap')
     stream.generate()
     stream.classify(IPS)
-    if mode == 2:   # do lebeling
+    if mode == 2:   # no lebeling
         stream.Group()
-    else:           # no labeling
+    else:           # do labeling
         stream.labelGroups()
+    print(f'Finished labeling for {name}...')
 
     # dump stream.json
     return make_record(name, stream, mode=mode)
@@ -163,7 +181,7 @@ def make_record(name, stream, *, mode):
         make_fingerprint(name, label, mode=mode)
 
     # dump stream.json
-    with open(f'./stream/{name}/stream.json', 'w') as json_file:
+    with open(make_path(f'stream/{name}/stream.json'), 'w') as json_file:
         json.dump(record, json_file)
     return record
 
@@ -172,17 +190,21 @@ def make_fingerprint(name, label, *, mode):
     """Make fingerprint."""
     if mode != 0:
         fp = fingerprintManager()
-        fp.GenerateAndUpdate(f'./stream/{name}/tmp', label)
+        fp.GenerateAndUpdate(make_path(f'stream/{name}/tmp'), label)
 
 
-def make_dataset(name, *, labels=None, overwrite=True, fingerprint=False):
+def make_dataset(name, labels=None, *, mode, overwrite=True, fingerprint=False):
     """Make dataset.
 
     Positional arguments:
         * name -- str, dataset source name
+        * labels -- dict, dataset labels
 
     Keyword arguments:
-        * labels -- dict, dataset labels
+        * mode -- int, preparation mode
+            |--> 0 -- stage 0, do labeling & no fingerprints
+            |--> 1 -- stage 1, do labeling & do fingerprints
+            |--> 2 -- stage 2, no labeling & do fingerprints
         * overwrite -- bool, if overwrite existing files
         * fingerprint -- bool, if generate and/or update fingerprints
 
@@ -190,33 +212,47 @@ def make_dataset(name, *, labels=None, overwrite=True, fingerprint=False):
         * dict -- dataset index
 
     """
+    print(f'Start making dataset for {name}...')
+
     # load JSON file
     if labels is None:
-        with open(f'./stream/{name}/strea.json', 'r') as file:
+        with open(make_path(f'stream/{name}/stream.json'), 'r') as file:
             labels = json.load(file)
 
+    fplist = list()
     for kind, group in labels.items():
         # only make dataset for type Background PC
         if kind != 'Background_PC':     continue
 
+        # make directory
+        pathlib.Path(make_path(f'dataset/{name}/{kind}/0')).mkdir(parents=True, exist_ok=True) # safe
+        pathlib.Path(make_path(f'dataset/{name}/{kind}/0')).mkdir(parents=True, exist_ok=True) # malicious
+
         # make fingerprints
         if fingerprint:
-            make_fingerprint(name, group, mode=-1)
+            make_fingerprint(name, group, mode=mode)
 
-        # make directory
-        pathlib.Path(f'./dataset/{name}/{kind}/0').mkdir(parents=True, exist_ok=True) # safe
-        pathlib.Path(f'./dataset/{name}/{kind}/1').mkdir(parents=True, exist_ok=True) # malicious
+        # identify figerprints
+        group_keys = group.keys()
+        if mode == 2:
+            fp = fingerprintManager()
+            fpreport = fp.Identify(make_path(f'stream/{name}/tmp'), group)
+            for ipua in fpreport['is_malicious']:
+                fplist += group[ipua]
+            group_keys = fpreport['new_app']
 
         # enumerate files
-        for files in group.values():
-            for file in files:
-                label = int(file['ismalicious'])
+        for ipua in group_keys:
+            for file in group[ipua]:
+                label = int(file['is_malicious'])
                 srcfile = file["filename"]
                 dataset = file['filename'].replace('.pcap', '.dat')
-                loads(f'./stream{name}/tmp/{srcfile}', f"./dataset/{name}/{kind}/{label}/{dataset}", remove=overwrite)
+                loads(make_path(f'stream/{name}/tmp/{srcfile}'),
+                        make_path(f"dataset/{name}/{kind}/{label}/{dataset}"), remove=overwrite)
+    print(f'Finished making dataset for {name}...')
 
     # dump index.json
-    return make_index()
+    return make_index(fp=fplist)
 
 
 def loads(fin, fout, *, remove):
@@ -243,31 +279,35 @@ def dumps(name, byte):
         file.write(byte)
 
 
-def make_index():
+def make_index(*, fp=None):
     """Dump index.json."""
     # initialise index
     index = {
-        'Browser_PC' : {0: list(), 1: list()},
-        'Background_PC' : {0: list(), 1: list()},
-        'Browser_Phone' : {0: list(), 1: list()},
-        'Background_Phone' : {0: list(), 1: list()},
-        'Suspicious' : {0: list(), 1: list()},
+        'Browser_PC' : {'0': list(), '1': list()},
+        'Background_PC' : {'0': list(), '1': list()},
+        'Browser_Phone' : {'0': list(), '1': list()},
+        'Background_Phone' : {'0': list(), '1': list()},
+        'Suspicious' : {'0': list(), '1': list()},
     }
 
     # walk dataset
-    for path in os.listdir('./dataset'):
+    for path in os.listdir(make_path('dataset')):
         for kind in index:
-            for root, _, files in os.walk(f'./dataset/{path}/{kind}/1'):
+            for root, _, files in os.walk(make_path(f'dataset/{path}/{kind}/1')):
                 for file in files:
                     if os.path.getsize(f'{root}/{file}'):
-                        index[kind][1].append(f'{root}/{file}')
-            for root, _, files in os.walk(f'./dataset/{path}/{kind}/0'):
+                        index[kind]['1'].append(f'{root}/{file}')
+            for root, _, files in os.walk(make_path(f'dataset/{path}/{kind}/0')):
                 for file in files:
                     if os.path.getsize(f'{root}/{file}'):
-                        index[kind][0].append(f'{root}/{file}')
+                        index[kind]['0'].append(f'{root}/{file}')
+
+    # fingerprint report
+    if fp is not None:
+        index['is_malicious'] = fp
 
     # dump index.json
-    with open('./dataset/index.json', 'w') as index_file:
+    with open(make_path('dataset/index.json'), 'w') as index_file:
         json.dump(index, index_file)
     return index
